update dependencies, move to min src 11, drop mr jar, update build plugins, fix build-time warnings
