+++ .github/workflows/maven.yml
@@ -23,7 +23,7 @@ jobs:
 
     strategy:
       matrix:
-        java_version: [ 11, 17-ea ]
+        java_version: [ 11, 17 ]
 
     steps:
     - name: Checkout for build
@@ -35,5 +35,8 @@ jobs:
       with:
         distribution: 'zulu'
         java-version: ${{ matrix.java_version }}
+        cache: maven
     - name: Verify
-      run: cd istack-commons && mvn -B -V -U -C -Pstaging,oss-release clean install org.glassfish.copyright:glassfish-copyright-maven-plugin:check -Dgpg.skip=true -Dittest=true
+      run: |
+        cd istack-commons
+        mvn -B -V -U -C -Pstaging,oss-release clean install org.glassfish.copyright:glassfish-copyright-maven-plugin:check -Dgpg.skip=true -Dittest=true -Dcopyright.ignoreyear=true
