+++ istack-commons/maven-plugin/src/main/java/com/sun/istack/maven/MetainfServicesCompilerMojo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1997, 2022 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Distribution License v. 1.0, which is available at
@@ -89,8 +89,8 @@ public class MetainfServicesCompilerMojo extends AbstractMojo {
     /**
      * The current repository/network configuration of Maven.
      *
-     * @parameter default-value="${repositorySystemSession}"
-     * @readonly
+     * parameter default-value="${repositorySystemSession}"
+     * readonly
      */
     private RepositorySystemSession repoSession;
