+++ istack-commons/runtime/src/main/java/com/sun/istack/SAXException2.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2021 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1997, 2022 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Distribution License v. 1.0, which is available at
@@ -19,6 +19,9 @@ import org.xml.sax.SAXException;
  * @since 2.0 FCS
  */
 public class SAXException2 extends SAXException {
+
+    private static final long serialVersionUID = -707119042406163844L;
+
     public SAXException2(String message) {
         super(message);
     }
