+++ istack-commons/runtime/src/main/java/com/sun/istack/SAXParseException2.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2021 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1997, 2022 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Distribution License v. 1.0, which is available at
@@ -20,6 +20,9 @@ import org.xml.sax.Locator;
  * @since 2.0 FCS
  */
 public class SAXParseException2 extends SAXParseException {
+
+    private static final long serialVersionUID = 1304853690034671001L;
+
     public SAXParseException2(String message, Locator locator) {
         super(message, locator);
     }
