+++ istack-commons/runtime/src/main/java/com/sun/istack/logging/StackHelper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2021 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2022 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Distribution License v. 1.0, which is available at
@@ -12,12 +12,11 @@ package com.sun.istack.logging;
 
 /**
  * Utils for stack trace analysis
- * 
- * @author Marek Potociar
- * @author Fabian Ritzmann
+ *
+ * @author Daniel Kec
  */
 class StackHelper {
-    
+
     /**
      * Function returns the name of the caller method for the method executing this
      * function.
@@ -25,27 +24,13 @@ class StackHelper {
      * @return caller method name from the call stack of the current {@link Thread}.
      */
     static String getCallerMethodName() {
-        return getStackMethodName(5);
-    }
-
-    /**
-     * Method returns the name of the method that is on the {@code methodIndexInStack}
-     * position in the call stack of the current {@link Thread}.
-     *
-     * @param methodIndexInStack index to the call stack to get the method name for.
-     * @return the name of the method that is on the {@code methodIndexInStack}
-     *         position in the call stack of the current {@link Thread}.
-     */
-    static String getStackMethodName(final int methodIndexInStack) {
-        final String methodName;
-
-        final StackTraceElement[] stack = Thread.currentThread().getStackTrace();
-        if (stack.length > methodIndexInStack + 1) {
-            methodName = stack[methodIndexInStack].getMethodName();
-        } else {
-            methodName = "UNKNOWN METHOD";
-        }
-
-        return methodName;
+        return StackWalker.getInstance(StackWalker.Option.RETAIN_CLASS_REFERENCE)
+                .walk(frames -> frames
+                        //Its a method of first declaring class after istack Logger class
+                        .dropWhile(f -> !Logger.class.equals(f.getDeclaringClass()))
+                        .dropWhile(f -> Logger.class.equals(f.getDeclaringClass()))
+                        .findFirst()
+                        .map(StackWalker.StackFrame::getMethodName)
+                        .orElse("UNKNOWN METHOD"));
     }
 }
