+++ istack-commons/soimp/src/main/java/com/sun/istack/soimp/ProcessingException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2018 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2022 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Distribution License v. 1.0, which is available at
@@ -14,6 +14,9 @@ package com.sun.istack.soimp;
  * @author Kohsuke Kawaguchi
  */
 public class ProcessingException extends Exception {
+
+    private static final long serialVersionUID = -6214429602282488762L;
+
     public ProcessingException(String message) {
         super(message);
     }
