+++ istack-commons/src/main/resources/copyright-exclude
