+++ istack-commons/tools/src/main/java/com/sun/istack/tools/Utils.java
@@ -10,52 +10,14 @@
 
 package com.sun.istack.tools;
 
-import java.lang.reflect.Field;
 import java.net.Authenticator;
-import java.security.AccessController;
-import java.security.PrivilegedAction;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
-/**
- *
- * @author lukas
- */
 final class Utils {
 
-    private static final Logger LOGGER = Logger.getLogger(DefaultAuthenticator.class.getName());
 
     private Utils() {}
 
     static Authenticator getCurrentAuthenticator() {
-        final Field f = getTheAuthenticator();
-        if (f == null) {
-            return null;
-        }
-
-        try {
-            AccessController.doPrivileged((PrivilegedAction<Void>) () -> {
-                f.setAccessible(true);
-                return null;
-            });
-            return (Authenticator) f.get(null);
-        } catch (IllegalAccessException | IllegalArgumentException ex) {
-            LOGGER.log(Level.FINEST, "Cannot get Authenticator instance", ex);
-            return null;
-        } finally {
-            AccessController.doPrivileged((PrivilegedAction<Void>) () -> {
-                f.setAccessible(false);
-                return null;
-            });
-        }
-    }
-
-    private static Field getTheAuthenticator() {
-        try {
-            return Authenticator.class.getDeclaredField("theAuthenticator");
-        } catch (NoSuchFieldException | SecurityException ex) {
-            LOGGER.log(Level.FINEST, "Cannot find Authenticator field", ex);
-            return null;
-        }
+        return Authenticator.getDefault();
     }
 }
